Benchmarks: Minor refactoring
Minor refactoring in response to PR criticism

- Subclass AbstractComponent instead of AbstractLifecycleComponent
- Move utility method out of AbstractBenchmarkService
